// libs/annuaire/shared/model/annuaire.ts
export interface Annuaire {
  identifiantDeLEtablissement: string;
  nomEtablissement: string;
  typeEtablissement: string;
  statutPublicPrive: string;
  adresse1: string;
  adresse2: string;
  adresse3: string;
  codePostal: string;
  codeCommune: string;
  nomCommune: string;
  codeDepartement: string;
  codeAcademie: string;
  codeRegion: string;
  ecoleMaternelle: number;
  ecoleElementaire: number;
  voieGenerale: string;
  voieTechnologique: string;
  voieProfessionnelle: string;
  telephone: string;
  fax: string;
  web: string;
  mail: string;
  restauration: number;
  hebergement: number;
  ulis: number;
  apprentissage: string;
  segpa: string;
  sectionArts: string;
  sectionCinema: string;
  sectionTheatre: string;
  sectionSport: string;
  sectionInternationale: string;
  sectionEuropeenne: string;
  lyceeAgricole: string;
  lyceeMilitaire: string;
  lyceeDesMetiers: string;
  postBac: string;
  appartenanceEducationPrioritaire: string;
  greta: string;
  sirenSiret: string;
  nombreDEleves: number;
  ficheOnisep: string;
  position: string;
  typeContratPrive: string;
  libelleDepartement: string;
  libelleAcademie: string;
  libelleRegion: string;
  coordxOrigine: number;
  coordyOrigine: number;
  epsgOrigine: string;
  nomCirconscription: string;
  latitude: number;
  longitude: number;
  precisionLocalisation: string;
  dateOuverture: string;
  dateMajLigne: string;
  etat: string;
  ministereTutelle: string;
  multiUai: number;
  rpiConcentre: number;
  rpiDisperse: string;
  codeNature: number;
  libelleNature: string;
  codeTypeContratPrive: string;
  pial: string;
  etablissementMere: string;
  typeRattachementEtablissementMere: string;
  codeCirconscription: string;
  codeZoneAnimationPedagogique: string;
  libelleZoneAnimationPedagogique: string;
  codeBassinFormation: string;
  libelleBassinFormation: string;
  anneeScolaire: string;
  multiRythme: string;
  lundiMatinDebut: string;
  lundiMatinFin: string;
  lundiApresMidiDebut: string;
  lundiApresMidiFin: string;
  lundi: string;
  mardiMatinDebut: string;
  mardiMatinFin: string;
  mardiApresMidiDebut: string;
  mardiApresMidiFin: string;
  mardi: string;
  mercrediMatinDebut: string;
  mercrediMatinFin: string;
  mercredi: string;
  jeudiMatinDebut: string;
  jeudiMatinFin: string;
  jeudiApresMidiDebut: string;
  jeudiApresMidiFin: string;
  jeudi: string;
  vendrediMatinDebut: string;
  vendrediMatinFin: string;
  vendrediApresMidiDebut: string;
  vendrediApresMidiFin: string;
  vendredi: string;
}